整理测试目录
