+++ base/lib/src/storage/storage.dart
@@ -99,11 +99,9 @@ class Storage {
   ///
   /// 使用注意：
   ///
-  /// 1、资源流需要在上传结束后自行关闭，SDK 内部不做关闭操作
+  /// 1、只支持分片上传
   ///
-  /// 2、只支持分片上传
-  ///
-  /// 3、如果客户端从[Stream]接收字节的速度大于上传速度会造成内存使用量上涨
+  /// 2、如果客户端从[Stream]接收字节的速度大于上传速度会造成内存使用量上涨
   ///
   /// [length] 资源字节长度
   ///
