+++ base/test/put/put_by_parts/put_file_test.dart
@@ -10,13 +10,13 @@ import 'package:qiniu_sdk_base/src/storage/methods/put/by_part/put_parts_task.da
 import 'package:qiniu_sdk_base/src/storage/resource/resource.dart';
 import 'package:test/test.dart';
 
-import '../config.dart';
-import 'put_controller_builder.dart';
+import '../../config.dart';
+import '../helpers.dart';
 
 void main() {
   configEnv();
 
-  test('putStream customVars should works well.', () async {
+  test('customVars&returnBody should works well.', () async {
     final storage = Storage();
 
     final auth = Auth(
@@ -38,37 +38,35 @@ void main() {
       'x:ext': 'testXExt',
     };
 
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final response = await storage.putStream(
-      file.openRead(),
+    var putController = PutController();
+    final response = await storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
       options: PutOptions(
-        key: 'test_for_put_parts.mp4',
+        key: fileKeyForPart,
         partSize: 1,
         customVars: customVars,
+        controller: putController,
       ),
     );
 
-    expect(response.key, 'test_for_put_parts.mp4');
+    expect(response.key, fileKeyForPart);
     expect(response.rawData['type'], 'testXType');
     expect(response.rawData['ext'], 'testXExt');
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream should works well.', () async {
+  test('putFile should works well.', () async {
     final storage = Storage();
     final pcb = PutControllerBuilder();
     var callnumber = 0;
     pcb.putController.addSendProgressListener((percent) {
       callnumber++;
     });
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final response = await storage.putStream(
-      file.openRead(),
+    final response = await storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
       options: PutOptions(
-        key: 'test_for_put_parts.mp4',
+        key: fileKeyForPart,
         partSize: 1,
         controller: pcb.putController,
       ),
@@ -80,85 +78,64 @@ void main() {
     pcb.testAll();
 
     // 不设置参数的情况
-    final responseNoOps = await storage.putStream(
-      file.openRead(),
+    final responseNoOps = await storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
     );
 
     expect(responseNoOps, isA<PutResponse>());
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream should throw error with incorrect partSize.', () async {
+  test('putFile should throw error with incorrect partSize.', () async {
     final storage = Storage();
-    final file = File('test_resource/test_for_put_parts.mp4');
     try {
-      await storage.putStream(
-        file.openRead(),
+      await storage.putFile(
+        fileForPart,
         token,
-        file.readAsBytesSync().length,
-        options: PutOptions(
-          key: 'test_for_put_parts.mp4',
-          partSize: 0,
-        ),
+        options: PutOptions(partSize: 0),
       );
     } catch (e) {
       expect(e, isA<AssertionError>());
     }
 
     try {
-      await storage.putStream(
-        file.openRead(),
+      await storage.putFile(
+        fileForPart,
         token,
-        file.readAsBytesSync().length,
-        options: PutOptions(
-          key: 'test_for_put_parts.mp4',
-          partSize: 1025,
-        ),
+        options: PutOptions(partSize: 1025),
       );
     } catch (e) {
       expect(e, isA<AssertionError>());
     }
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream should works well while response 612.', () async {
+  test('putFile should works well while response 612.', () async {
     final httpAdapterTest = HttpAdapterTestWith612();
     final storage = Storage(config: Config(httpClientAdapter: httpAdapterTest));
-    final file = File('test_resource/test_for_put_parts.mp4');
+    final response = await storage.putFile(
+      fileForPart,
+      token,
+      options: PutOptions(key: fileKeyForPart, partSize: 1),
+    );
 
-    try {
-      await storage.putStream(
-        file.openRead(),
-        token,
-        file.readAsBytesSync().length,
-        options: PutOptions(
-          key: 'test_for_put_parts.mp4',
-          partSize: 1,
-        ),
-      );
-    } catch (e) {
-      /// httpAdapterTest 应该会触发一次 612 response
-      expect(httpAdapterTest.completePartsTaskResponse612, true);
-      expect(e, isA<StorageError>());
-      expect((e as StorageError).code, 612);
-    }
+    /// httpAdapterTest 应该会触发一次 612 response
+    expect(httpAdapterTest.completePartsTaskResponse612, true);
+    expect(response, isA<PutResponse>());
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream can be cancelled.', () async {
+  test('putFile can be cancelled.', () async {
     final pcb = PutControllerBuilder();
     final storage = Storage(config: Config(hostProvider: HostProviderTest()));
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final key = 'test_for_put_parts.mp4';
+    final key = fileKeyForPart;
     pcb.putController.addSendProgressListener((percent) {
       // 开始上传并且 InitPartsTask 设置完缓存后取消
       if (percent > 0.1) {
         pcb.putController.cancel();
       }
     });
-    var future = storage.putStream(
-      file.openRead(),
+    var future = storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
       options: PutOptions(
         key: key,
         partSize: 1,
@@ -181,10 +158,9 @@ void main() {
     ]);
 
     try {
-      await storage.putStream(
-        file.openRead(),
+      await storage.putFile(
+        fileForPart,
         token,
-        file.readAsBytesSync().length,
         options: PutOptions(
           key: key,
           partSize: 1,
@@ -199,15 +175,15 @@ void main() {
 
     expect(future, throwsA(isA<StorageError>()));
 
-    final response = await storage.putBytes(
-      file.readAsBytesSync(),
+    final response = await storage.putFile(
+      fileForPart,
       token,
       options: PutOptions(key: key, partSize: 1),
     );
     expect(response, isA<PutResponse>());
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream can be resumed.', () async {
+  test('putFile can be resumed.', () async {
     final storage = Storage(config: Config(hostProvider: HostProviderTest()));
     final putController = PutController();
     putController.addSendProgressListener((percent) {
@@ -217,12 +193,11 @@ void main() {
       }
     });
 
-    final future = storage.putStream(
-      File('test_resource/test_for_put_parts.mp4').openRead(),
+    final future = storage.putFile(
+      fileForPart,
       token,
-      File('test_resource/test_for_put_parts.mp4').readAsBytesSync().length,
       options: PutOptions(
-        key: 'test_for_put_parts.mp4',
+        key: fileKeyForPart,
         partSize: 1,
         controller: putController,
       ),
@@ -237,62 +212,22 @@ void main() {
 
     expect(future, throwsA(TypeMatcher<StorageError>()));
 
-    final response = await storage.putBytes(
-      File('test_resource/test_for_put_parts.mp4').readAsBytesSync(),
+    final response = await storage.putFile(
+      fileForPart,
       token,
-      options: PutOptions(key: 'test_for_put_parts.mp4', partSize: 1),
+      options: PutOptions(key: fileKeyForPart, partSize: 1),
     );
 
     expect(response, isA<PutResponse>());
   }, skip: !isSensitiveDataDefined);
 
-  test('putstream should throw error if there is a same task is working.',
-      () async {
-    final storage = Storage();
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final key = 'test_for_put_parts.mp4';
-
-    var errorOccurred = false;
-
-    // 故意不 await，让后面发送一个相同的任务
-    // ignore: unawaited_futures
-    storage.putStream(
-      file.openRead(),
-      token,
-      file.lengthSync(),
-      id: 'test',
-      options: PutOptions(
-        key: key,
-        partSize: 1,
-      ),
-    );
-
-    try {
-      await storage.putStream(
-        file.openRead(),
-        token,
-        file.lengthSync(),
-        id: 'test',
-        options: PutOptions(
-          key: key,
-          partSize: 1,
-        ),
-      );
-    } catch (e) {
-      errorOccurred = true;
-      expect(e, isA<StorageError>());
-      expect((e as StorageError).type, StorageErrorType.IN_PROGRESS);
-    }
-    expect(errorOccurred, true);
-  }, skip: !isSensitiveDataDefined);
-
-  test('putStream should works well with cacheProvider.', () async {
+  test('putFile should works well with cacheProvider.', () async {
     final cacheProvider = CacheProviderForTest();
     final config = Config(cacheProvider: cacheProvider);
     final storage = Storage(config: config);
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final key = 'test_for_put_parts.mp4';
-    final resource = FileResource(file: file, length: file.lengthSync());
+    final key = fileKeyForPart;
+    final resource =
+        FileResource(file: fileForPart, length: fileForPart.lengthSync());
 
     /// 手动初始化一个初始化文件的任务，确定分片上传的第一步会被缓存
     final task = InitPartsTask(
@@ -314,15 +249,10 @@ void main() {
       }
     });
 
-    final future = storage.putStream(
-      file.openRead(),
+    final future = storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
-      options: PutOptions(
-        key: key,
-        partSize: 1,
-        controller: putController,
-      ),
+      options: PutOptions(key: key, partSize: 1, controller: putController),
     );
 
     /// 这个时候应该只缓存了初始化的缓存信息
@@ -350,14 +280,10 @@ void main() {
     await cacheProvider.clear();
     cacheProvider.callNumber = 0;
 
-    final response = await storage.putStream(
-      file.openRead(),
+    final response = await storage.putFile(
+      fileForPart,
       token,
-      file.readAsBytesSync().length,
-      options: PutOptions(
-        key: key,
-        partSize: 1,
-      ),
+      options: PutOptions(key: key, partSize: 1),
     );
 
     expect(response, isA<PutResponse>());
@@ -368,16 +294,50 @@ void main() {
     expect(cacheProvider.callNumber, 3);
   }, skip: !isSensitiveDataDefined);
 
-  test('putStream\'s status and progress should works well.', () async {
+  test('putFile should throw error if there is a same task is working.',
+      () async {
+    final storage = Storage();
+    final key = fileKeyForPart;
+
+    var errorOccurred = false;
+
+    // 故意不 await，让后面发送一个相同的任务
+    // ignore: unawaited_futures
+    storage.putFile(
+      fileForPart,
+      token,
+      options: PutOptions(
+        key: key,
+        partSize: 1,
+      ),
+    );
+
+    try {
+      await storage.putFile(
+        fileForPart,
+        token,
+        options: PutOptions(
+          key: key,
+          partSize: 1,
+        ),
+      );
+    } catch (e) {
+      errorOccurred = true;
+      expect(e, isA<StorageError>());
+      expect((e as StorageError).type, StorageErrorType.IN_PROGRESS);
+    }
+    expect(errorOccurred, true);
+  }, skip: !isSensitiveDataDefined);
+
+  test('putFile\'s status and progress should works well.', () async {
     final storage = Storage();
     final pcb = PutControllerBuilder();
 
-    final response = await storage.putStream(
-      File('test_resource/test_for_put_parts.mp4').openRead(),
+    final response = await storage.putFile(
+      fileForPart,
       token,
-      File('test_resource/test_for_put_parts.mp4').readAsBytesSync().length,
       options: PutOptions(
-        key: 'test_for_put_parts.mp4',
+        key: fileKeyForPart,
         partSize: 1,
         controller: pcb.putController,
       ),
@@ -387,57 +347,4 @@ void main() {
       ..testProcess()
       ..testStatus(targetProgressList: [0.001, 0.002, 0.99, 1]);
   }, skip: !isSensitiveDataDefined);
-}
-
-class HttpAdapterTestWith612 extends HttpClientAdapter {
-  /// 记录 CompletePartsTask 被创建的次数
-  /// 第一次我们拦截并返回 612，第二次不拦截
-  bool completePartsTaskResponse612 = false;
-  final DefaultHttpClientAdapter _adapter = DefaultHttpClientAdapter();
-  @override
-  void close({bool force = false}) {
-    _adapter.close(force: force);
-  }
-
-  @override
-  Future<ResponseBody> fetch(RequestOptions options,
-      Stream<Uint8List>? requestStream, Future? cancelFuture) async {
-    /// 如果是 CompletePartsTask 发出去的请求，则返回 612
-    if (options.path.contains('uploads/') &&
-        options.method == 'POST' &&
-        !completePartsTaskResponse612) {
-      completePartsTaskResponse612 = true;
-      return ResponseBody.fromString('', 612);
-    }
-    return _adapter.fetch(options, requestStream, cancelFuture);
-  }
-}
-
-class HostProviderTest extends HostProvider {
-  @override
-  Future<String> getUpHost({
-    required String accessKey,
-    required String bucket,
-  }) async {
-    // token 中 bucket 对应的地区
-    return 'https://upload-na0.qiniup.com';
-  }
-
-  @override
-  void freezeHost(String host) {}
-
-  @override
-  bool isFrozen(String host) {
-    return false;
-  }
-}
-
-class CacheProviderForTest extends DefaultCacheProvider {
-  int callNumber = 0;
-  @override
-  // ignore: unnecessary_overrides
-  Future setItem(String key, String item) {
-    callNumber++;
-    return super.setItem(key, item);
-  }
 }
