--- base/test/put/put_controller_test.dart
@@ -1,16 +1,0 @@
-import 'package:qiniu_sdk_base/src/storage/storage.dart';
-import 'package:test/test.dart';
-
-import '../config.dart';
-
-void main() {
-  configEnv();
-  test('put controller basic.', () async {
-    final controller = PutController();
-    expect(controller.isCancelled, false);
-    controller.cancel();
-    expect(controller.isCancelled, true);
-    controller.cancel();
-    expect(controller.isCancelled, true);
-  }, skip: !isSensitiveDataDefined);
-}
